Split codex session modules
